+++ vite.config.js
@@ -2,5 +2,6 @@ import { defineConfig } from "vite";
 
 export default defineConfig({
     base: "/jspsych-typing/",
+    build: {target: "esnext"}
   });